Refacgtoring some more, moving the entities into a separate file
